dropdown menu to select organism and store the user selected organism
